+++ app/upgrades.go
@@ -3,17 +3,27 @@ package app
 import (
 	"context"
 	"fmt"
+	"slices"
 
 	storetypes "cosmossdk.io/store/types"
 	upgradetypes "cosmossdk.io/x/upgrade/types"
 	"github.com/cosmos/cosmos-sdk/codec"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
-	clientkeeper "github.com/cosmos/ibc-go/v8/modules/core/02-client/keeper"
 )
 
-func (app *ChainApp) RegisterUpgradeHandlers(cdc codec.BinaryCodec, clientKeeper clientkeeper.Keeper) {
+func (app *ChainApp) RegisterUpgradeHandlers(cdc codec.BinaryCodec) {
 	planName := "v5.0"
 	app.UpgradeKeeper.SetUpgradeHandler(planName, func(ctx context.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+		sdkCtx := sdk.UnwrapSDKContext(ctx)
+		{
+			params := app.ICAHostKeeper.GetParams(sdkCtx)
+			msg := "/ibc.applications.interchain_accounts.host.v1.MsgModuleQuerySafe"
+			if !slices.Contains(params.AllowMessages, msg) {
+				params.AllowMessages = append(params.AllowMessages, msg)
+				app.ICAHostKeeper.SetParams(sdkCtx, params)
+			}
+		}
 		return app.ModuleManager.RunMigrations(ctx, app.configurator, fromVM)
 	})